Working on improve loading buttons

+++ src/components/Dialogs/RemoveCompanyDialog/RemoveCompanyDialog.js
@@ -9,8 +9,10 @@ import { useState, useEffect } from "react";
 import { Company } from "../../../api";
 import { useAuth } from "../../../hooks";
 import { NotificationSnackbar } from "../../NotificationSnackbar";
-import CircularProgress from "@mui/material/CircularProgress";
 import { LoadingButton } from "@mui/lab";
+import WarningRoundedIcon from "@mui/icons-material/WarningRounded";
+import { companyStatus } from "../../../utils";
+import { ErrorDialog } from "../ErrorDialog";
 
 export function RemoveCompanyDialog({
   selectedCompany,
@@ -39,6 +41,16 @@ export function RemoveCompanyDialog({
   const handleAccept = async () => {
     const companyController = new Company();
     try {
+      if (selectedCompany.status === companyStatus.DELETED) {
+        return (
+          <ErrorDialog
+            errorMessage={"La empresa ya fue eliminada anteriormente."}
+            openDialog={openDialog}
+            onDialogOpenChange={onDialogOpenChange}
+          />
+        );
+      }
+
       setLoading(true); // Inicia la carga
 
       await companyController.delete(accessToken, selectedCompany.id);
@@ -73,6 +85,12 @@ export function RemoveCompanyDialog({
                 {`¿Desea eliminar la empresa ${selectedCompany.businessName}?`}
               </Typography>
               <Typography variant="body1" paragraph>
+                <WarningRoundedIcon
+                  sx={{
+                    verticalAlign: "middle", // Alinea verticalmente con el texto
+                    marginRight: "4px", // Opcional: agrega margen a la derecha para separar el icono del texto
+                  }}
+                />
                 Ten en cuenta que también se eliminarán los depósitos asociados.
               </Typography>
             </>
